Add end to end test
Executes built binary to ensure that any regressions in the complete CLI
tool will be caught by the tests (e.g. changes to the command line
arguments/commands), and the behaviour of the actual tool around files
on disk.

In the style of
https://github.com/rustworkshop/gitopolis/blob/989381071fac589fc4f41f4da03ffde9320a3213/tests/end_to_end_tests.rs

https://github.com/timabell/disk-hog-backup/issues/10

This patch was coaxed out of windsurf ai IDE with lots of prompting